Intégration de la carte dans le parcours MPRA
