--- components/MarFinder.tsx
@@ -1,11 +1,0 @@
-export default function MARFinder() {
-  return (
-    <div>
-      En fonction de votre commune,{' '}
-      <a href="https://france-renov.gouv.fr/annuaire-ar/recherche">
-        découvrez la liste des accompagnateurs rénov' proches
-      </a>
-      .
-    </div>
-  )
-}
